# docs/openapi.json
{
  "components": {
    "schemas": {
      "Body_tai_len_api_video_post": {
        "properties": {
          "blur": {
            "default": "auto",
            "title": "Blur",
            "type": "string"
          },
          "blur_box": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blur Box"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "font_scale": {
            "default": 0.42,
            "title": "Font Scale",
            "type": "number"
          },
          "force": {
            "default": false,
            "title": "Force",
            "type": "boolean"
          },
          "force_asr": {
            "default": false,
            "title": "Force Asr",
            "type": "boolean"
          },
          "lang": {
            "default": "en",
            "title": "Lang",
            "type": "string"
          },
          "model": {
            "default": "large-v3",
            "title": "Model",
            "type": "string"
          },
          "model_dich": {
            "default": "nllb-cuc-bo",
            "title": "Model Dich",
            "type": "string"
          },
          "nhom": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nhom"
          },
          "separate": {
            "default": false,
            "title": "Separate",
            "type": "boolean"
          },
          "vad": {
            "default": true,
            "title": "Vad",
            "type": "boolean"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_tai_len_api_video_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Dich phu de video Anh -> Viet",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/cong-viec": {
      "get": {
        "operationId": "danh_sach_cong_viec_api_cong_viec_get",
        "parameters": [
          {
            "in": "query",
            "name": "trang_thai",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Trang Thai"
            }
          },
          {
            "in": "query",
            "name": "sap_xep",
            "required": false,
            "schema": {
              "default": "tao_luc",
              "title": "Sap Xep",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "thu_tu",
            "required": false,
            "schema": {
              "default": "desc",
              "title": "Thu Tu",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "trang",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Trang",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "moi_trang",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Moi Trang",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Danh Sach Cong Viec Api Cong Viec Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Danh Sach Cong Viec"
      }
    },
    "/api/cong-viec/{cid}": {
      "get": {
        "operationId": "trang_thai_api_cong_viec__cid__get",
        "parameters": [
          {
            "in": "path",
            "name": "cid",
            "required": true,
            "schema": {
              "title": "Cid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Trang Thai Api Cong Viec  Cid  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trang Thai"
      }
    },
    "/api/cong-viec/{cid}/hop": {
      "post": {
        "description": "Nhan hop da ve roi chay tiep tu buoc dich. Bo qua ve = blur off.",
        "operationId": "nhan_hop_api_cong_viec__cid__hop_post",
        "parameters": [
          {
            "in": "path",
            "name": "cid",
            "required": true,
            "schema": {
              "title": "Cid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Than",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Nhan Hop Api Cong Viec  Cid  Hop Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Nhan Hop"
      }
    },
    "/api/cong-viec/{cid}/ket-qua": {
      "get": {
        "operationId": "ket_qua_api_cong_viec__cid__ket_qua_get",
        "parameters": [
          {
            "in": "path",
            "name": "cid",
            "required": true,
            "schema": {
              "title": "Cid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ket Qua"
      }
    },
    "/api/cong-viec/{cid}/khung": {
      "get": {
        "description": "Danh sach khung mau kem moc thoi gian va cau thoai de frontend hien len.",
        "operationId": "khung_api_cong_viec__cid__khung_get",
        "parameters": [
          {
            "in": "path",
            "name": "cid",
            "required": true,
            "schema": {
              "title": "Cid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Khung Api Cong Viec  Cid  Khung Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Khung"
      }
    },
    "/api/cong-viec/{cid}/khung/{i}": {
      "get": {
        "operationId": "mot_khung_api_cong_viec__cid__khung__i__get",
        "parameters": [
          {
            "in": "path",
            "name": "cid",
            "required": true,
            "schema": {
              "title": "Cid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "i",
            "required": true,
            "schema": {
              "title": "I",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mot Khung"
      }
    },
    "/api/lich-su": {
      "get": {
        "description": "Ai doi gi trong nhom/thuat ngu (do trigger CSDL ghi, nen khong duong ghi nao bo sot).",
        "operationId": "lich_su_api_lich_su_get",
        "parameters": [
          {
            "in": "query",
            "name": "nhom",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Nhom"
            }
          },
          {
            "in": "query",
            "name": "doi_tuong",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Doi Tuong"
            }
          },
          {
            "in": "query",
            "name": "hanh_dong",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hanh Dong"
            }
          },
          {
            "in": "query",
            "name": "trang",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Trang",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "moi_trang",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Moi Trang",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Lich Su Api Lich Su Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lich Su"
      }
    },
    "/api/nhat-ky": {
      "get": {
        "operationId": "nhat_ky_api_nhat_ky_get",
        "parameters": [
          {
            "in": "query",
            "name": "video_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Video Id"
            }
          },
          {
            "in": "query",
            "name": "buoc",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Buoc"
            }
          },
          {
            "in": "query",
            "name": "ket_qua",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ket Qua"
            }
          },
          {
            "in": "query",
            "name": "trang",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Trang",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "moi_trang",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Moi Trang",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Nhat Ky Api Nhat Ky Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Nhat Ky"
      }
    },
...
    "/api/video": {
      "post": {
        "operationId": "tai_len_api_video_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_tai_len_api_video_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Tai Len Api Video Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tai Len"
      }
    }
  }
}